updated programming guide examples in line with changes to the guide for int64_t

+++ example/prog-guide/async-distobj-reduce_to_rank0.hpp
@@ -1,17 +1,17 @@
-int reduce_to_rank0(int my_hits) {
+int64_t reduce_to_rank0(int64_t my_hits) {
     // initialize this rank's part of the distributed object with the local value
-    upcxx::dist_object<int> all_hits(my_hits);
-    int hits = 0;
+    upcxx::dist_object<int64_t> all_hits(my_hits);
+    int64_t hits = 0;
     // rank 0 gets all the values asynchronously
     if (upcxx::rank_me() == 0) {
-        upcxx::future<int> f = upcxx::make_future(my_hits);
+        upcxx::future<int64_t> f = upcxx::make_future(my_hits);
         for (int i = 1; i < upcxx::rank_n(); i++) {
             // get the future value from remote rank i
-            upcxx::future<int> remote_rank_val = fetch(all_hits, i);
+            upcxx::future<int64_t> remote_rank_val = fetch(all_hits, i);
             // create a future that combines f and the remote rank's result
-            upcxx::future<int, int> combined_f = upcxx::when_all(f, remote_rank_val);
+            upcxx::future<int64_t, int64_t> combined_f = upcxx::when_all(f, remote_rank_val);
             // get the future for the combined result, summing the values
-            f = combined_f.then([](int a, int b) { return a + b; });
+            f = combined_f.then([](int64_t a, int64_t b) { return a + b; });
         }
         // wait for the chain to complete
         hits = f.wait();
